// src/integrations/cms/generated/cms-schema.d.ts
export interface components {
    schemas: {
// ...
        ApiPromotionPromotionDocument: {
// ...
            content: unknown[];
// ...
        };
// ...
        SharedLinkEntry: {
            /** @description A string field */
            label?: string;
            /** @description A string field */
            url?: string;
            /**
             * @description A boolean field
             * @default false
             */
            external: boolean | null;
            /** @description An integer field */
            order?: number;
            /** @description A media field */
            icon?: components["schemas"]["PluginUploadFileDocument"];
        };
        SharedButtonEntry: {
            /** @description A string field */
            label?: string;
            /** @description A string field */
            url?: string;
            /**
             * @description An enum field
             * @default primary
             * @enum {string}
             */
            style: "primary" | "secondary" | "link";
            /** @description An integer field */
            order?: number;
            /** @description A media field */
            icon?: components["schemas"]["PluginUploadFileDocument"];
            /** @description A media field */
            iconhover?: components["schemas"]["PluginUploadFileDocument"];
        };
// ...
        ApiNavigationItemNavigationItemDocument: {
            /**
             * Format: uuid
             * @description The document ID, represented by a UUID
             */
            documentId: string;
            id: string | number;
            /** @description A string field */
            name?: string;
            /** @description A string field */
            title?: string;
            /** @description A string field */
            url?: string;
            /**
             * @description An integer field
             * @default 0
             */
            order: number;
            /** @description A datetime field */
            createdAt?: string;
            /** @description A datetime field */
            updatedAt?: string;
            /**
             * @description A datetime field
             * @default 2026-06-04T06:49:03.732Z
             */
            publishedAt: string;
            /** @description A media field */
            icon?: components["schemas"]["PluginUploadFileDocument"];
            /** @description A relational field */
            navigations?: components["schemas"]["ApiNavigationNavigationDocument"][];
            /** @description A relational field */
            parent?: components["schemas"]["ApiNavigationItemNavigationItemDocument"];
            /** @description A relational field */
            children?: components["schemas"]["ApiNavigationItemNavigationItemDocument"][];
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
// ...
    "navigation/get/navigations_by_id": {
        parameters: {
            query?: {
                fields?: ("name" | "type_menu" | "on_off" | "createdAt" | "updatedAt" | "publishedAt")[];
                populate?: "*" | ("navigation_items" | "iconshare" | "ButtonShare") | ("navigation_items" | "iconshare" | "ButtonShare")[];
                filters?: {
                    [key: string]: unknown;
                };
                sort?: ("name" | "type_menu" | "on_off" | "createdAt" | "updatedAt" | "publishedAt") | ("name" | "type_menu" | "on_off" | "createdAt" | "updatedAt" | "publishedAt")[] | {
                    [key: string]: "asc" | "desc";
                } | {
                    [key: string]: "asc" | "desc";
                }[];
                status?: "draft" | "published";
                hasPublishedVersion?: boolean | ("true" | "false");
            };
            header?: never;
            path: {
                id: string;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description OK */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": {
                        data: {
                            /**
                             * Format: uuid
                             * @description The document ID, represented by a UUID
                             */
                            documentId: string;
                            id: string | number;
                            /** @description A string field */
                            name: string;
                            /**
                             * @description An enum field
                             * @enum {string}
                             */
                            type_menu: "header" | "footer" | "mobile" | "h-khcc" | "f-khcc";
                            /**
                             * @description A boolean field
                             * @default true
                             */
                            on_off: boolean | null;
                            /** @description A datetime field */
                            createdAt?: string;
                            /** @description A datetime field */
                            updatedAt?: string;
                            /**
                             * @description A datetime field
                             * @default 2026-06-04T06:49:03.744Z
                             */
                            publishedAt: string;
                            /** @description A relational field */
                            navigation_items?: components["schemas"]["ApiNavigationItemNavigationItemDocument"][];
                            /** @description A component field */
                            iconshare?: components["schemas"]["SharedLinkEntry"][];
                            /** @description A component field */
                            ButtonShare?: components["schemas"]["SharedButtonEntry"][];
                        };
                    };
                };
            };
            /** @description Bad request */
            400: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
            /** @description Unauthorized */
            401: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
            /** @description Forbidden */
            403: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
            /** @description Not found */
            404: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
            /** @description Internal server error */
            500: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
        };
    };
// ...
}